Should preserve the compilation database link
The link is also needed by YouCompleteMe.

+++ reformat.sh
@@ -1,15 +1,14 @@
 SECONDS=0
 
 # Create the symbolic link we need.
-ln -s build/compile_commands.json
+if [ ! -f $COMPILATION_DATABASE ]; then
+  ln -s build/compile_commands.json
+fi
 
 echo Reformatting...
 clang-format -i walls-of-doom/*.[hc]
 
 echo Analyzing...
 clang-tidy walls-of-doom/*.[c]
-
-# Clean up.
-rm compile_commands.json
 
 echo "Done after $SECONDS seconds."
